requirements.txt
gradio
qdrant-client
PyMuPDF
python-docx
google-generativeai
google-genai
python-dotenv
tqdm
agno
duckduckgo-search
yfinance
pypdf